# api/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "version": "v0.1-beta",
        "title": "Story Protocol API Reference"
    },
    "paths": {
...
        "/api/v1/policyframeworks/{pfwmId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a PolicyFrameworkManager",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PolicyFrameworkManagers"
                ],
                "summary": "Get a PolicyFrameworkManager",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "PolicyFrameworkManager ID",
                        "name": "pfwmId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.PolicyFrameworkManagerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/policymanagers": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of PolicyFrameworkManagers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PolicyFrameworkManagers"
                ],
                "summary": "List PolicyFrameworkManagers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.PolicyFrameworkManagerRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.PolicyFrameworkManagersResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/royalties": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of Royalties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Royalties"
                ],
                "summary": "List Royalties",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.RoyaltyRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.RoyaltiesResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/royalties/{royaltyId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a Royalty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Royalties"
                ],
                "summary": "Get a Royalty",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Royalty ID",
                        "name": "royaltyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.RoyaltyResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/royaltypays": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of RoyaltyPays",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RoyaltyPays"
                ],
                "summary": "List RoyaltyPays",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.RoyaltyPayRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.RoyaltyPaysResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/royaltypays/{royaltyPayId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a RoyaltyPay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RoyaltyPays"
                ],
                "summary": "Get a RoyaltyPay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RoyaltyPay ID",
                        "name": "royaltyPayId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.RoyaltyPayResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "betav0.PFWMQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object",
                    "properties": {
                        "address": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "betav0.PolicyFrameworkManager": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "licenseUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "betav0.PolicyFrameworkManagerRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.PFWMQueryOptions"
                }
            }
        },
        "betav0.PolicyFrameworkManagerResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.PolicyFrameworkManager"
                }
            }
        },
        "betav0.PolicyFrameworkManagersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.PolicyFrameworkManager"
                    }
                }
            }
        },
...
        "betav0.RoyaltiesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.Royalty"
                    }
                }
            }
        },
        "betav0.Royalty": {
            "type": "object",
            "properties": {
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ipId": {
                    "type": "string"
                },
                "royaltyPolicy": {
                    "type": "string"
                }
            }
        },
        "betav0.RoyaltyPay": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payerIpId": {
                    "type": "string"
                },
                "receiverIpId": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "betav0.RoyaltyPayQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object",
                    "properties": {
                        "ipId": {
                            "type": "string"
                        },
                        "payerIpId": {
                            "type": "string"
                        },
                        "receiverIpId": {
                            "type": "string"
                        },
                        "sender": {
                            "type": "string"
                        },
                        "token": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "betav0.RoyaltyPayRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.RoyaltyPayQueryOptions"
                }
            }
        },
        "betav0.RoyaltyPayResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.RoyaltyPay"
                }
            }
        },
        "betav0.RoyaltyPaysResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.RoyaltyPay"
                    }
                }
            }
        },
        "betav0.RoyaltyQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object",
                    "properties": {
                        "ipId": {
                            "type": "string"
                        },
                        "royaltyPolicy": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "betav0.RoyaltyRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.RoyaltyQueryOptions"
                }
            }
        },
        "betav0.RoyaltyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.Royalty"
                }
            }
        },
...
    }
}